\name{rgam-package}
\alias{rgam-package}
\docType{package}
\title{Package \sQuote{rgam}}

\description{
  The \code{rgam} package is an implementation of algorithms for outlier-robust
  fit for Generalized Additive Models described in \cite{Azadeh and
    Salibian-Barrera (2011)}. It has one entry function,
  \code{\link{rgam}}. For more details refer to the relevant help files.
}

\author{
Matias Salibian-Barrera and Davor Cubranic.

Maintainer: Davor Cubranic <cubranic@stat.ubc.ca>
}

\references{
  Azadeh, A. and Salibian-Barrera, M. (2011).
  \emph{An outlier-robust fit for Generalized Additive Models with
    applications to disease outbreak detection.}
  To appear in the Journal of the American Statistical Association.
}

\keyword{ package }
\keyword{ robust }

\examples{
## see examples for rgam and ili.visits.
}
